backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
langdetect==1.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.0
anthropic==0.9.0
google-generativeai==0.3.2
requests==2.31.0
python-dateutil==2.8.2
